pyproject.toml
[project]
name = "local-brain"
version = "0.2.0"
description = "Skill-based LLM tasks using local Ollama models with tool calling"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Ismael Martinez", email = "ismaelmartinez@gmail.com"}
]
keywords = ["ollama", "llm", "code-review", "cli", "tool-calling", "agents"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "ollama>=0.6.1",
    "pyyaml>=6.0.0",
    "jinja2>=3.0.0",
    "click>=8.0.0",
]

[project.scripts]
local-brain = "local_brain.cli:main"

[project.urls]
Homepage = "https://github.com/IsmaelMartinez/local-brain"
Repository = "https://github.com/IsmaelMartinez/local-brain"
Documentation = "https://github.com/IsmaelMartinez/local-brain#readme"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
